Switch to uncontrolled inputs for inspector
React was already a little with unhappy with us switching between
controlled (in design mode) and uncontrolled (in preflight mode).
Instead, we always run in uncontrolled mode, which also fixes a really
subtle but annoying issue where React keeps struggling to "undo" our inputs
because the input was controlled, so the color picker didn't animate
smoothly.

Also includes a fix for the focused element not getting overwritten.